Refactor blog page to fast refresh friendly system
