+++ re_pages/Blog_Article.js
@@ -1,6 +1,6 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
-import * as Blog from "../pages/blog.js";
+import * as Blog from "./Blog.js";
 import * as Icon from "../components/Icon.js";
 import * as Meta from "../components/Meta.js";
 import * as Util from "../common/Util.js";
